prompt optimize updated
